Redid forEach Loop/Added stylings
Removed all document.write() from the forEach loop and replaced it with a new function that creates list items for each pokemon in the repository. These list items are then displayed as buttons on the main index page.

Also added stylings to all buttons, and when clicked on they now write the Pokemon's name to the console.